[RenderComponent + Queue port, and cleanups]
Ported over the components and classes responsible for rendering.
Render components haven't been tested yet, but once MeshRendererComponent and UIRendererComponent are ported it will be quick to test.

Cleaned up some of the warnings that kept cropping up when I compiled (which just made it annoying to sift through for the real errors).
Added some markup to the readme to make it easier to read.

+++ source/ObjectAllocator.cpp
@@ -18,11 +18,11 @@ void puni::ObjectAllocator::addToDestroyQueue(Object * obj)
 
 void puni::ObjectAllocator::processDestroyQueue()
 {
-	for (int i = 0; i < destroyQueue.size(); ++i)
+	for (unsigned long i = 0; i < destroyQueue.size(); ++i)
 	{
 		if (destroyQueue[i] != nullptr)
 		{
-			for (int k = 0; k < instanceList.size(); ++k)
+			for (unsigned long k = 0; k < instanceList.size(); ++k)
 			{
 				if (instanceList[k]->id == destroyQueue[i]->id)
 				{
@@ -46,7 +46,7 @@ puni::ObjectAllocator::ObjectAllocator()
 
 puni::ObjectAllocator::~ObjectAllocator()
 {
-	for (int i = 0; i < instanceList.size(); ++i)
+	for (unsigned long i = 0; i < instanceList.size(); ++i)
 	{
 		delete instanceList[i];
 	}
